cancel/close modal

+++ src/index.js
@@ -1,3 +1,4 @@
+import { openModal } from "./assets/javascripts/modal";
 import "./assets/styles/styles.scss";
 import "./index.scss";
 
@@ -69,21 +70,27 @@ const createArticles = () => {
   });
   deleteButtons.forEach((button) => {
     button.addEventListener("click", async (event) => {
-      try {
-        const target = event.target;
-        const articleId = target.dataset.id;
-        const response = await fetch(
-          `https://restapi.fr/api/article/${articleId}`,
-          {
-            method: "DELETE",
-          }
-        );
-        const body = await response.json();
-        console.log(body);
-        fetchArticle();
-      } catch (e) {
-        console.log("e : ", e);
+      const result = await openModal(
+        "Etes vous sûr de vouloir supprimer votre article ?"
+      );
+      if (result === true) {
+        try {
+          const target = event.target;
+          const articleId = target.dataset.id;
+          const response = await fetch(
+            `https://restapi.fr/api/article/${articleId}`,
+            {
+              method: "DELETE",
+            }
+          );
+          const body = await response.json();
+          console.log(body);
+          fetchArticle();
+        } catch (e) {
+          console.log("e : ", e);
+        }
       }
+      console.log(result);
     });
   });
 };
